refactor: move CH fork snapshot rewrite into CH package
