+++ lib/forkvm/README.md
@@ -4,24 +4,17 @@ This package contains low-level helpers used by instance forking.
 
 ## Why this package exists
 
-Forking is mostly filesystem and snapshot identity rewriting work. Keeping that logic outside `lib/instances` keeps lifecycle orchestration focused on state transitions and locking.
+Forking includes filesystem cloning work that is independent of any specific hypervisor. Keeping that logic outside `lib/instances` keeps lifecycle orchestration focused on state transitions and locking.
 
 ## What it provides
 
 - `CopyGuestDirectory(srcDir, dstDir)`
   - Recursively copies a guest directory.
-  - Skips runtime sockets (for example `ch.sock`) because they are process-local artifacts.
-- `RewriteSnapshotConfig(configPath, opts)`
-  - Rewrites `snapshots/snapshot-latest/config.json` for a forked cloud-hypervisor instance.
-  - Supports:
-    - source data-dir -> target data-dir path remap
-    - vsock CID/socket rewrite
-    - serial log path rewrite
-    - network fields (`tap`, `ip`, `mac`, `mask`) rewrite
+  - Skips runtime sockets because they are process-local artifacts.
 
 ## How it is used
 
-`lib/instances/fork.go` calls this package to clone guest data and prepare copied snapshot state. `lib/instances/restore.go` uses it when a standby fork restores with a fresh network identity.
+`lib/instances/fork.go` calls this package to clone guest data before metadata and hypervisor-specific fork preparation.
 
 ## Safety notes
