+++ lib/hypervisor/cloudhypervisor/fork.go
@@ -3,7 +3,6 @@ package cloudhypervisor
 import (
 	"context"
 
-	"github.com/kernel/hypeman/lib/forkvm"
 	"github.com/kernel/hypeman/lib/hypervisor"
 )
 
@@ -15,22 +14,5 @@ func (s *Starter) PrepareFork(ctx context.Context, req hypervisor.ForkPrepareReq
 		return nil
 	}
 
-	var netCfg *forkvm.SnapshotNetworkConfig
-	if req.Network != nil {
-		netCfg = &forkvm.SnapshotNetworkConfig{
-			TAPDevice: req.Network.TAPDevice,
-			IP:        req.Network.IP,
-			MAC:       req.Network.MAC,
-			Netmask:   req.Network.Netmask,
-		}
-	}
-
-	return forkvm.RewriteSnapshotConfig(req.SnapshotConfigPath, forkvm.SnapshotRewriteOptions{
-		SourceDataDir: req.SourceDataDir,
-		TargetDataDir: req.TargetDataDir,
-		VsockCID:      &req.VsockCID,
-		VsockSocket:   req.VsockSocket,
-		SerialLogPath: req.SerialLogPath,
-		Network:       netCfg,
-	})
+	return rewriteSnapshotConfigForFork(req.SnapshotConfigPath, req)
 }
